pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.2.15",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
